Fix delete all to not produce sql in statement
Conflicts:
	activerecord/test/cases/associations/has_many_associations_test.rb

Merge pull request #14727 from robin850/patch-17

Add a changelog entry for #14546 [ci skip]

Conflicts:
	activerecord/CHANGELOG.md

Conflicts:
	activerecord/CHANGELOG.md
	activerecord/lib/active_record/associations/collection_association.rb
	activerecord/test/cases/associations/has_many_associations_test.rb
	activerecord/test/cases/test_case.rb

+++ activerecord/lib/active_record/associations/collection_association.rb
@@ -226,7 +226,7 @@ module ActiveRecord
             ActiveRecord::Base.logger ? ActiveRecord::Base.logger.warn(message) : $stderr.puts(message)
           end
 
-          if loaded? || dependent == :destroy
+          if (loaded? || dependent == :destroy) && dependent != :delete_all
             delete_or_destroy(load_target, dependent)
           else
             delete_records(:all, dependent)
